+++ tests/tests_pytorch/deprecated_api/test_remove_2-0.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """Test deprecated functionality which will be removed in v2.0.0."""
-from unittest import mock
 
 import numpy
 import pytest
@@ -45,7 +44,6 @@ from pytorch_lightning.utilities.device_parser import (
 )
 from pytorch_lightning.utilities.optimizer import optimizer_to_device, optimizers_to_device
 from pytorch_lightning.utilities.seed import pl_worker_init_function, reset_seed, seed_everything
-from pytorch_lightning.utilities.xla_device import inner_f, pl_multi_process, XLADeviceUtils
 
 
 def test_v1_10_deprecated_on_colab_kaggle_func():
@@ -59,23 +57,6 @@ def test_v1_10_deprecated_device_dtype_module_mixin():
 
     with pytest.deprecated_call(match="mixins.DeviceDtypeModuleMixin` has been deprecated in v1.8.0"):
         MyModule()
-
-
-def test_v1_10_deprecated_xla_device_utilities():
-    with pytest.deprecated_call(match="xla_device.inner_f` has been deprecated in v1.8.0"):
-        inner_f(mock.Mock(), mock.Mock())
-
-    with pytest.deprecated_call(match="xla_device.pl_multi_process` has been deprecated in v1.8.0"):
-        pl_multi_process(mock.Mock)
-
-    with pytest.deprecated_call(match="xla_device.XLADeviceUtils` has been deprecated in v1.8.0"):
-        XLADeviceUtils()
-
-    with pytest.deprecated_call(match="xla_device.XLADeviceUtils.xla_available` has been deprecated in v1.8.0"):
-        XLADeviceUtils.xla_available()
-
-    with pytest.deprecated_call(match="xla_device.XLADeviceUtils.tpu_device_exists` has been deprecated in v1.8.0"):
-        XLADeviceUtils.tpu_device_exists()
 
 
 def test_v1_10_deprecated_apply_func_utilities():